make scripts

+++ database/birthday/delete_by_name.sql
@@ -1,3 +1,5 @@
 DELETE FROM public.birthday
-WHERE "name" = $1
+WHERE
+    "name" = $1
+    AND user_id = $1
 RETURNING *
